cleanup for 2023

+++ src/AdventOfCode.Solutions2020/Day11/Solution.cs
@@ -48,8 +48,8 @@ public class Solution : ISolution
             _maxY = InternalMap.Max(m => m.Key.Y);
         }
 
-        private readonly int _maxX;
-        private readonly int _maxY;
+        private readonly long _maxX;
+        private readonly long _maxY;
             
         public int UpdateSeats()
         {
